requirements.txt
pytest==7.2.2
selenium==4.8.2
webdriver-manager==3.8.5
allure-pytest==2.12.0

